+++ packages/core/build.config.ts
@@ -4,7 +4,8 @@ export default defineBuildConfig({
   entries: [
     './src/index',
     './src/client',
-    './src/types'
+    './src/types',
+    './src/worker'
   ],
   clean: true,
   declaration: true,